factor out common mapping code from doobie/sql test
